updates
I did not see any errors in your code. I would have liked to see the information centered instead of off to the left. I also thought that the blue and green made it unprofessional since it was a portfolio and business card. But I thought that the content that you included was good and sounded professional